# docs/api/openapi-fase0.json
{
  "components": {
    "schemas": {
      "AjusteCreado": {
        "description": "La respuesta del alta: lo mismo que la fila, más el nivel de alerta en\nque quedó el producto (lo deriva el servidor, que es la única autoridad\ndel umbral — decisión 2).",
        "properties": {
          "aplicado_por": {
            "title": "Aplicado Por",
            "type": "string"
          },
          "cantidad": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cantidad"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "delta": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Delta",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "motivo": {
            "title": "Motivo",
            "type": "string"
          },
          "nivel": {
            "title": "Nivel",
            "type": "string"
          },
          "producto_id": {
            "format": "uuid",
            "title": "Producto Id",
            "type": "string"
          },
          "stock_contado": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stock Contado"
          },
          "stock_resultante": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Stock Resultante",
            "type": "string"
          },
          "tipo": {
            "title": "Tipo",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tipo",
          "producto_id",
          "delta",
          "motivo",
          "aplicado_por",
          "stock_resultante",
          "nivel"
        ],
        "title": "AjusteCreado",
        "type": "object"
      },
      "AjusteCrear": {
        "additionalProperties": false,
        "description": "Un ajuste por conteo o una merma. ONLINE obligatorio (ADR-020): el\ndelta se calcula contra el stock del servidor en el momento.\n\n`id` es REQUERIDO (decisión 4): es la PK de `ajustes_inventario` y la\núnica ancla que hace seguro el reintento de una merma, que es un delta\nrelativo. El `motivo` es obligatorio: un ajuste sin justificación es un\ndesfalco con buenos modales.",
        "properties": {
          "cantidad": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 99999999999.999,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cantidad"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "motivo": {
            "maxLength": 300,
            "minLength": 3,
            "title": "Motivo",
            "type": "string"
          },
          "producto_id": {
            "format": "uuid",
            "title": "Producto Id",
            "type": "string"
          },
          "stock_contado": {
            "anyOf": [
              {
                "maximum": 99999999999.999,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stock Contado"
          },
          "tipo": {
            "enum": [
              "ajuste",
              "merma"
            ],
            "title": "Tipo",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tipo",
          "producto_id",
          "motivo"
        ],
        "title": "AjusteCrear",
        "type": "object"
      },
...
      "CompraCrear": {
        "additionalProperties": false,
        "description": "Una compra a proveedor. `proveedor_nombre` es texto libre (ADR-020:\nla factura es un papel; no hay tabla de proveedores). El `id` del\ncliente se acepta como PK (ADR-017): reenviar la misma compra es un\nno-op. El total NO viaja: lo calcula el servidor (decisión 7).",
        "properties": {
          "fecha": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha"
          },
          "id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CompraItemEntrada"
            },
            "maxItems": 200,
            "minItems": 1,
            "title": "Items",
            "type": "array"
          },
          "observaciones": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observaciones"
          },
          "proveedor_nombre": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Proveedor Nombre",
            "type": "string"
          }
        },
        "required": [
          "proveedor_nombre",
          "items"
        ],
        "title": "CompraCrear",
        "type": "object"
      },
      "CompraDetalleSalida": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "fecha": {
            "format": "date",
            "title": "Fecha",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CompraItemSalida"
            },
            "title": "Items",
            "type": "array"
          },
          "observaciones": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observaciones"
          },
          "proveedor_nombre": {
            "title": "Proveedor Nombre",
            "type": "string"
          },
          "total_centavos": {
            "title": "Total Centavos",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "proveedor_nombre",
          "fecha",
          "total_centavos",
          "items"
        ],
        "title": "CompraDetalleSalida",
        "type": "object"
      },
      "CompraItemEntrada": {
        "additionalProperties": false,
        "description": "Una línea de factura. El costo es el de ESTA compra: al confirmarse\nactualiza `ultimo_costo` del producto (ADR-020).",
        "properties": {
          "cantidad": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 99999999999.999,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Cantidad"
          },
          "costo_unitario_centavos": {
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Costo Unitario Centavos",
            "type": "integer"
          },
          "producto_id": {
            "format": "uuid",
            "title": "Producto Id",
            "type": "string"
          }
        },
        "required": [
          "producto_id",
          "cantidad",
          "costo_unitario_centavos"
        ],
        "title": "CompraItemEntrada",
        "type": "object"
      },
...
      "CompraSalida": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "fecha": {
            "format": "date",
            "title": "Fecha",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "observaciones": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observaciones"
          },
          "proveedor_nombre": {
            "title": "Proveedor Nombre",
            "type": "string"
          },
          "total_centavos": {
            "title": "Total Centavos",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "proveedor_nombre",
          "fecha",
          "total_centavos"
        ],
        "title": "CompraSalida",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "default": "ERROR",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": false,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "PagedList_AjusteSalida_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AjusteSalida"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "skip": {
            "title": "Skip",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "skip",
          "limit"
        ],
        "title": "PagedList[AjusteSalida]",
        "type": "object"
      },
      "PagedList_CompraSalida_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CompraSalida"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "skip": {
            "title": "Skip",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "skip",
          "limit"
        ],
        "title": "PagedList[CompraSalida]",
        "type": "object"
      },
...
      "PagedList_StockSalida_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StockSalida"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "skip": {
            "title": "Skip",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "skip",
          "limit"
        ],
        "title": "PagedList[StockSalida]",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "\nAPI regional de Vendi. Fase 1: fundación + catálogo + ventas con sync offline.\n\nEl negocio (tenant) se resuelve del claim `organization` del token: el alias de\nla Organization de Keycloak **es** el identificador del negocio. Un usuario que\npertenece a varios negocios indica cuál con la cabecera `X-Tenant-Id`, y solo\npuede elegir entre los que ya trae su token.\n",
    "title": "Vendi API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/compras": {
      "get": {
        "operationId": "listar_compras_api_v1_compras_get",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedList_CompraSalida_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Falta el token o no es válido"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Falta el permiso o el negocio está suspendido"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request malformado (validación de estructura o de dominio)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Listar compras",
        "tags": [
          "inventario"
        ]
      },
      "post": {
        "description": "En la MISMA transacción: la compra, sus ítems, un movimiento `compra`\npor línea, `stock_actual` y `ultimo_costo` de cada producto, y el evento\n`compra.registrada` (ADR-020). Acepta el `id` que traiga el cliente\n(ADR-017): reenviar la misma compra devuelve la existente, sin duplicar\nfila, stock ni evento. El total lo calcula el servidor por línea; el\n`proveedor_nombre` es texto libre (la factura es un papel: no hay tabla\nde proveedores).",
        "operationId": "registrar_compra_api_v1_compras_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompraCrear"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDetalleSalida"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Falta el token o no es válido"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Falta el permiso o el negocio está suspendido"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "El id de la compra ya existe (en este u otro negocio)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request malformado (validación de estructura o de dominio)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Registrar una compra a proveedor",
        "tags": [
          "inventario"
        ]
      }
    },
    "/api/v1/compras/{compra_id}": {
      "get": {
        "operationId": "ver_compra_api_v1_compras__compra_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "compra_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Compra Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDetalleSalida"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Falta el token o no es válido"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Falta el permiso o el negocio está suspendido"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "La compra no existe"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request malformado (validación de estructura o de dominio)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ver una compra con sus ítems",
        "tags": [
          "inventario"
        ]
      }
    },
...
    "/api/v1/inventario/ajustes": {
      "get": {
        "description": "La auditoría del «¿quién movió el arroz?»: cada fila lleva su motivo\ny quién la aplicó.",
        "operationId": "listar_ajustes_api_v1_inventario_ajustes_get",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedList_AjusteSalida_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Falta el token o no es válido"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Falta el permiso o el negocio está suspendido"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request malformado (validación de estructura o de dominio)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Listar ajustes y mermas",
        "tags": [
          "inventario"
        ]
      },
      "post": {
        "description": "La única operación de inventario que EXIGE conexión (ADR-020): el\ndelta se calcula contra el stock del servidor en el momento del conteo —\nun ajuste offline corrompería el contador de forma no conmutativa. El\n`motivo` es obligatorio. El `id` del cliente es requerido y ancla la\nidempotencia: el reintento idéntico devuelve lo ya respondido sin mover\nstock; el divergente es 409 `ajuste_id_divergente`. Un conteo que cuadra\n(delta 0) graba la fila pero no escribe movimiento en el libro.",
        "operationId": "registrar_ajuste_api_v1_inventario_ajustes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AjusteCrear"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AjusteCreado"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Falta el token o no es válido"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Falta el permiso o el negocio está suspendido"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "El id del ajuste ya existe con datos distintos"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request malformado (validación de estructura o de dominio)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Registrar un ajuste por conteo o una merma (ONLINE)",
        "tags": [
          "inventario"
        ]
      }
    },
    "/api/v1/inventario/stock": {
      "get": {
        "description": "El nivel lo deriva el servidor con la misma función que dispara\n`inventario.alerta_stock`: una sola definición del umbral. El stock\nnegativo es un dato legítimo (ADR-020) y viaja como `agotado`.",
        "operationId": "estado_de_stock_api_v1_inventario_stock_get",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Solo productos agotados o por debajo del mínimo",
            "in": "query",
            "name": "solo_alertas",
            "required": false,
            "schema": {
              "default": false,
              "description": "Solo productos agotados o por debajo del mínimo",
              "title": "Solo Alertas",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedList_StockSalida_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Falta el token o no es válido"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Falta el permiso o el negocio está suspendido"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request malformado (validación de estructura o de dominio)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Estado de stock con su nivel (agotado/crítico/bajo/ok)",
        "tags": [
          "inventario"
        ]
      }
    },
...
  }
}